% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http_client.R
\name{eva.authenticate}
\alias{eva.authenticate}
\title{Authenticate the session (usually not called manually)}
\usage{
eva.authenticate(session)
}
\arguments{
\item{session}{}
}
\description{
Authenticate the session (usually not called manually)
}
