% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http_client.R
\name{eva.call}
\alias{eva.call}
\title{Perform a RPC call on HMI service}
\usage{
eva.call(session, method, params = list())
}
\arguments{
\item{session}{}

\item{method}{RPC method}

\item{params}{method params}
}
\description{
Perform a RPC call on HMI service
}
