% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http_client.R
\name{eva.credentials}
\alias{eva.credentials}
\title{Set session credentials for login/password authentication}
\usage{
eva.credentials(session, user, password)
}
\arguments{
\item{session}{}

\item{user}{user login}

\item{password}{user password}
}
\description{
Set session credentials for login/password authentication
}
