% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history.R
\name{eva.history.append_oid}
\alias{eva.history.append_oid}
\title{Append item OID mapping to a request}
\usage{
eva.history.append_oid(
  request,
  oid,
  status = FALSE,
  value = FALSE,
  database = NULL,
  xopts = NULL
)
}
\arguments{
\item{request}{}

\item{oid}{item OID}

\item{status}{status mapping: TRUE or col name}

\item{value}{value mapping: TRUE or col name}

\item{database}{database service to load state from}

\item{xopts}{extra database request options}
}
\description{
Append item OID mapping to a request
}
