% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history.R
\name{eva.history.push}
\alias{eva.history.push}
\title{Push data to a database service}
\usage{
eva.history.push(session, request, data, database = "default")
}
\arguments{
\item{session}{}

\item{request}{request object}

\item{data}{file path or a data frame}

\item{database}{database to push data to (default: default)}
}
\description{
Push data to a database service
}
