% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history.R
\name{eva.history.request}
\alias{eva.history.request}
\title{Create a new item state history request}
\usage{
eva.history.request(params_csv = NULL)
}
\arguments{
\item{params_csv}{Load oid map params from a CSV file}
}
\description{
Create a new item state history request
}
