% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http_client.R
\name{eva.session}
\alias{eva.session}
\title{Create a new EVA ICS session}
\usage{
eva.session(
  url = "http://localhost:7727",
  user = NULL,
  password = NULL,
  key = NULL,
  timeout = 120,
  mlkit = FALSE
)
}
\arguments{
\item{url}{HMI or front-end server URL}

\item{user}{user login (for login/password authentication)}

\item{password}{user password}

\item{key}{API key (for API key authentication)}

\item{timeout}{timeout for HTTP requests}

\item{mlkit}{if ML Kit server is installed: TRUE (if front-end server is used) or ML Kit server URL}
}
\description{
Create a new EVA ICS session
}
